Arreglo doublewritable

+++ Practica2/Hadoop/DevSTDReducer.java
@@ -9,7 +9,7 @@ import org.apache.hadoop.mapreduce.Reducer;
 public class DevSTDReducer<Key> extends Reducer<Key,DoubleWritable, Key,DoubleWritable> {
 
 
-   public void reduce(Key key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException {
+   public void reduce(Key key, Iterable<DoubleWritable> values, Context context) throws IOException, InterruptedException {
 
      int sum = 0;
      int field = Integer.parseInt(key.toString());
